requirements.txt
# Core Flask dependencies
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8
gunicorn==21.2.0

# Firebase dependencies
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0

# API utilities
Flask-Mail==0.9.1
flask_cors==4.0.0
slowapi==0.1.0
openai==1.12.0

# Additional utilities
python-dotenv==0.19.0
requests==2.26.0
markdown==3.6.0
beautifulsoup4==4.12.3
googlemaps==4.10.0

# Required dependencies
certifi==2024.12.14
charset-normalizer==2.0.12
idna==3.10
urllib3==1.26.20
googleapis-common-protos==1.66.0
googlemaps==4.10.0
googletrans==4.0.2
grpcio==1.69.0
grpcio-status==1.48.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.11
nest-asyncio==1.6.0
packaging==24.2
preshed==3.0.9
proto-plus==1.25.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==0.19.0
requests==2.26.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
starlette==0.41.3
thinc==8.3.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.1
zipp==3.21.0
gunicorn==21.2.0
slowapi
firebase_admin
Flask-Mail==0.9.1
markdown
openai
flask_cors
